+++ CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Creates a password and prints it at the cursor location, possibly replacing
+  the selected text. If there are many cursors, create a password for each of
+  them.
+
 ## Fixed
 
 - A lot of little bugfixing